pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# Project settings -----------------------------------------------------------------------------------------------------
[project]
name = "ultralytics_extension"
dynamic = ["version"]
authors = [
    {name = "cvteam"}  # Correctly formatted as an object
]

# Required dependencies ------------------------------------------------------------------------------------------------
dependencies = [
    "numpy>=1.23.0",
    "matplotlib>=3.3.0",
    "opencv-python>=4.6.0",
    "pillow>=7.1.2",
    "pyyaml>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "torch==2.5.1+cu124",
    "torchvision==0.20.1+cu124",
#    "torch>=1.8.0",
#    "torch>=1.8.0,!=2.4.0; sys_platform == 'win32'", # Windows CPU errors w/ 2.4.0 https://github.com/ultralytics/ultralytics/issues/15049
    "torchvision>=0.9.0",
    "tqdm>=4.64.0", # progress bars
    "psutil", # system utilization
    "py-cpuinfo", # display CPU info
    "pandas>=1.1.4",
    "seaborn>=0.11.0", # plotting
    "ultralytics-thop>=2.0.0", # FLOPs computation https://github.com/ultralytics/thop
    "albumentations>2.0.0", 
]


# Optional dependencies ------------------------------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "ipython",
    "pytest",
    "notebook",
    "jupyterlab",
    "ipykernel",
]

# Tools settings -------------------------------------------------------------------------------------------------------
[tool.setuptools.packages]  # configuration specific to the `setuptools` build backend.
# where = ["src"]
find = { where = ["src"], include = ["ultralytics", "ultralytics_extension"] }
